requirements.txt
aiosmtplib==4.0.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.1
daphne==4.1.2
diff-match-patch==20241021
Django==5.1.6
django-allauth==65.4.1
django-cacheops==7.1
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-etc==1.4.0
django-filter==24.3
django-hitcount==1.3.5
django-import-export==4.3.7
django-json-widget==2.0.1
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.2
django-simple-history==3.8.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-api-key==2.3.0
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
environs==14.1.1
funcy==2.0
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.4.2
marshmallow==3.26.1
msgpack==1.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
